% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ascii.R
\name{image_ascii}
\alias{image_ascii}
\title{ASCII art}
\usage{
image_ascii(
  image,
  chars = c("B", "S", "#", "&", "@", "€", "\%", "*", "!", ":", ".")
)
}
\arguments{
\item{image}{The path of the file to be processed}

\item{chars}{Character}
}
\value{
Character
}
\description{
ASCII art
}
\examples{
system.file("extdata", "R_logo.svg", package = "inkscaper") \%>\%
inx_svg2png() \%>\% magick::image_read() \%>\% image_ascii() \%>\% print()
}
