% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_actions.R
\name{inx_actions}
\alias{inx_actions}
\title{Title}
\usage{
inx_actions(input, actions, ext)
}
\arguments{
\item{input}{svg}

\item{actions}{actions from list}

\item{ext}{file output extension}
}
\value{
file for conversion
}
\description{
Title
}
\examples{
input = system.file("extdata", "MyStar.svg", package = "inkscaper", mustWork = TRUE)
grid::grid.newpage()
input \%>\% inx_actions(actions = NA, ext = ".png") \%>\%
png::readPNG() \%>\%
grid::grid.raster()
grid::grid.newpage()
input \%>\% inx_actions(actions = "select-by-id:MyStar;object-flip-vertical", ext = ".png") \%>\%
png::readPNG() \%>\%
grid::grid.raster()
}
