% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_actions_list.R
\name{inx_actions_list}
\alias{inx_actions_list}
\title{Get a list of Inkscape actions and put them in a tibble}
\usage{
inx_actions_list()
}
\value{
tibble with action names and descriptions
}
\description{
Get a list of Inkscape actions and put them in a tibble
}
\examples{
inx_actions_list()
}
