% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_extension.R
\name{inx_extension}
\alias{inx_extension}
\title{Title}
\usage{
inx_extension(input, inkscape_extension_name, ext)
}
\arguments{
\item{input}{file path}

\item{ext}{- extension of the output file}

\item{name}{of the inkscape_extension_name}
}
\value{
a ouput file
}
\description{
Title
}
\examples{
library(tidyverse)
"https://dev.w3.org/SVG/tools/svgweb/samples/svg-files/italian-flag.svg" \%>\%
inx_extension("lorem_ipsum.py", ".svg") \%>\%
inx_actions(actions="", ".png") \%>\% png::readPNG() -> img
ggplot() +
annotation_raster(img, -Inf, Inf, -Inf, Inf)
}
