% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_extensions_list.R
\name{inx_extensions_list}
\alias{inx_extensions_list}
\title{Build a dataframe with the names and descriptions of the Inkscape extensions}
\usage{
inx_extensions_list()
}
\value{
dataframe with the names and descriptions of the Inkscape extensions
}
\description{
Build a dataframe with the names and descriptions of the Inkscape extensions
}
\examples{
inx_extensions_list()
}
