% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_plot.R
\name{inx_plot}
\alias{inx_plot}
\title{Function to send plot produced with `ggplot()` to Inkscape window.
Works only on Desktop.}
\usage{
inx_plot(plot)
}
\arguments{
\item{plot}{produced with `ggplot()`}
}
\description{
Function to send plot produced with `ggplot()` to Inkscape window.
Works only on Desktop.
}
\examples{
\dontrun{
library(ggplot2)
{ggplot(iris,
aes(x = Petal.Length,
    y = Petal.Width,
    colour = Sepal.Length,
    shape = Species)) +
    geom_point()} \%>\%
inx_plot()
}
}
