% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_sf2svg.R
\name{inx_sf2svg}
\alias{inx_sf2svg}
\title{Title}
\usage{
inx_sf2svg(geom)
}
\arguments{
\item{geom}{simple features}
}
\value{
file name
}
\description{
Title
}
\examples{
library(tidyverse)
library(sf)
filename <- system.file("shape/nc.shp", package="sf")
nc <- st_read(filename)
inx_sf2svg(nc)
}
