% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_source.R
\name{inx_source}
\alias{inx_source}
\title{Title}
\usage{
inx_source(input)
}
\arguments{
\item{input}{svg file}
}
\value{
svg source xml
}
\description{
Title
}
\examples{
input = system.file("extdata", "R_logo.svg", package = "inkscaper", mustWork = TRUE)
input \%>\% inx_source()
}
