% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_style.R
\name{inx_style}
\alias{inx_style}
\title{Title}
\usage{
inx_style(input, id, style)
}
\arguments{
\item{input}{SVG file}

\item{id}{element id}

\item{style}{new style}
}
\value{
svg file
}
\description{
Title
}
\examples{
library(tidyverse)
# plot the original file
system.file("extdata", "rectangle.svg", package = "inkscaper") \%>\%
inx_actions(actions = NA, ext = ".png") \%>\%
png::readPNG(native = TRUE) \%>\%
grid::rasterGrob(interpolate=TRUE) -> img
ggplot() +
annotation_custom(img, xmin=-Inf, xmax=Inf, ymin=-Inf, ymax=Inf)
# modify style and plot
system.file("extdata", "rectangle.svg", package = "inkscaper") \%>\%
inx_style(id = "rect788", style = "fill:#FFFFFF;stroke:#000000;stroke-width:1") \%>\%
inx_actions(actions = NA, ext = ".png") \%>\%
png::readPNG(native = TRUE) \%>\%
grid::rasterGrob(interpolate=TRUE) -> img
ggplot() +
annotation_custom(img, xmin=-Inf, xmax=Inf, ymin=-Inf, ymax=Inf)
}
