% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_svg2sf.R
\name{inx_svg2sf}
\alias{inx_svg2sf}
\title{Title}
\usage{
inx_svg2sf(input)
}
\arguments{
\item{input}{file or url of the SVG}
}
\value{
simple features
}
\description{
Title
}
\examples{
library(tidyverse)
library(sf)
input = system.file("extdata", "R_logo.svg", package = "inkscaper", mustWork = TRUE)
input \%>\%
inx_svg2sf() \%>\%
ggplot() +
  geom_sf()

"https://dev.w3.org/SVG/tools/svgweb/samples/svg-files/tiger.svg" \%>\%
inx_svg2sf() \%>\%
ggplot() +
  geom_sf()
}
