% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_verbs_list.R
\name{inx_verbs_list}
\alias{inx_verbs_list}
\title{Build a dataframe with the names and descriptions of the Inkscape verbs}
\usage{
inx_verbs_list()
}
\value{
dataframe with the names and descriptions of the Inkscape verbs
}
\description{
Build a dataframe with the names and descriptions of the Inkscape verbs
}
\examples{
\dontrun{
inx_verbs_list()
}
}
