% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_version.R
\name{inx_version}
\alias{inx_version}
\title{Write the file in your working directory}
\usage{
inx_version()
}
\value{
Logical
}
\description{
Write the file in your working directory
}
\examples{
inx_version()
}
