% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_write.R
\name{inx_write}
\alias{inx_write}
\title{Title}
\usage{
inx_write(input, output)
}
\arguments{
\item{input}{Path of the file to be copied}
}
\value{
Name of the file in Working Directory
}
\description{
Title
}
\examples{
input = system.file("extdata", "R_logo.svg", package = "inkscaper")
inx_write(input, "my_file.svg")
}
