% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_to_dxf.R
\name{svg_to_dxf}
\alias{svg_to_dxf}
\title{Title}
\usage{
svg_to_dxf(input)
}
\arguments{
\item{input}{file or url of the SVG}
}
\value{
file name
}
\description{
Title
}
\examples{
library(tidyverse)
library(sf)
input = system.file("extdata", "R_logo.svg", package = "inkscaper", mustWork = TRUE)
input \%>\%
svg_to_dxf() \%>\%
st_read() \%>\%
ggplot() +
  geom_sf()

"https://dev.w3.org/SVG/tools/svgweb/samples/svg-files/tiger.svg" \%>\%
svg_to_dxf() \%>\%
st_read() \%>\%
ggplot() +
  geom_sf()
}
