% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_actions.R
\name{inx_actions}
\alias{inx_actions}
\title{Title}
\usage{
inx_actions(input, actions, ext)
}
\arguments{
\item{input}{svg}

\item{actions}{actions from list}

\item{ext}{file output extension}
}
\value{
file for conversion
}
\description{
Title
}
\examples{
library(tidyverse)
library(sf)
# empty plot - attempt to convert SVG containing native {rect} objects
"https://dev.w3.org/SVG/tools/svgweb/samples/svg-files/italian-flag.svg" \%>\%
svg_to_dxf() \%>\%
st_read() \%>\% ggplot() +
  geom_sf()
# right plot - attempt to convert SVG containing paths from native {rect} objects
"https://dev.w3.org/SVG/tools/svgweb/samples/svg-files/italian-flag.svg" \%>\%
inx_actions(actions = list('select-all', 'object-to-path'), ext = ".svg") \%>\%
svg_to_dxf() \%>\%
st_read() \%>\% ggplot() +
  geom_sf()
}
