% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_actions.R
\name{inx_actions}
\alias{inx_actions}
\title{Run Inkscape actions}
\usage{
inx_actions(input, actions, ext)
}
\arguments{
\item{input}{Path or url of the SVG input file.}

\item{actions}{vector of actions to be performed. Check vignetes for details.}

\item{ext}{file format of the output (ex. ".svg", ".png", ...)}
}
\value{
path of the outcome
}
\description{
Run Inkscape actions
}
\examples{
library(ggplot2)
img = "https://upload.wikimedia.org/wikipedia/commons/3/30/Den_Haag_wapen.svg" \%>\%
inx_actions(actions = NA, ext = ".png") \%>\%
png::readPNG(native = TRUE) \%>\%
grid::rasterGrob(interpolate=TRUE)
ggplot() +
annotation_custom(img, xmin=-Inf, xmax=Inf, ymin=-Inf, ymax=Inf)
}
