% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_extension.R
\name{inx_extension}
\alias{inx_extension}
\title{Run Inkscape extension}
\usage{
inx_extension(input, inkscape_extension_name, ext)
}
\arguments{
\item{input}{The path of the input file}

\item{inkscape_extension_name}{The name of the Inkscape extension}

\item{ext}{The extension of the output file}
}
\value{
The output file
}
\description{
Run Inkscape extension
}
\examples{
library(tidyverse)
"https://dev.w3.org/SVG/tools/svgweb/samples/svg-files/italian-flag.svg" \%>\%
inx_extension("lorem_ipsum.py", ".svg") \%>\%
inx_actions(actions="", ".png") \%>\% png::readPNG() -> img
ggplot() +
annotation_raster(img, -Inf, Inf, -Inf, Inf)
}
