% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_svg2sf.R
\name{inx_svg2sf}
\alias{inx_svg2sf}
\title{Title}
\usage{
inx_svg2sf(input)
}
\arguments{
\item{input}{Path or url of the SVG input file.}
}
\value{
Simple feature collection of geometry type LINESTRING.
}
\description{
Title
}
\examples{
library(ggplot2)
library(sf)
sf_from_svg <- "https://upload.wikimedia.org/wikipedia/commons/3/30/Den_Haag_wapen.svg" \%>\%
inx_svg2sf()
sf_from_svg
sf_from_svg \%>\%
ggplot() +
geom_sf()
}
