% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inx_write.R
\name{inx_write}
\alias{inx_write}
\title{Write the file in your working directory}
\usage{
inx_write(input, output)
}
\arguments{
\item{input}{The path of the file to be copied}

\item{output}{The name of the file in the working directory}
}
\value{
Logical
}
\description{
Write the file in your working directory
}
\examples{
input = system.file("extdata", "R_logo.svg", package = "inkscaper")
inx_write(input, "my_file.svg")
}
